+++ src/main/java/frc/robot/commands/AlignToScoringLocationCommand.java
@@ -3,6 +3,7 @@ package frc.robot.commands;
 import java.util.function.Supplier;
 
 import edu.wpi.first.math.MathUtil;
+import edu.wpi.first.math.controller.PIDController;
 import edu.wpi.first.math.geometry.Pose2d;
 import frc.robot.commands.supplier.EstimatedPoseSupplier;
 import frc.robot.commands.supplier.ScoringLocationSupplier;
@@ -10,8 +11,9 @@ import frc.robot.filters.DriveInput;
 import frc.robot.pose.AlignmentCalculator;
 import frc.robot.pose.ScoringLocation;
 import frc.robot.subsystems.DriveSubsystem;
+import frc.robot.util.StoppingCounter;
 
-
+import static frc.robot.commands.DriveCommandConstants.*;
 /**
  * Tries to align to the target scoring location, by rotating the robot about its axis
  * (Operator can still drive in the meantime)
@@ -21,8 +23,12 @@ import frc.robot.subsystems.DriveSubsystem;
  * 
  * @author dcowden
  */
-public class AlignToScoringLocationCommand extends BaseDrivePIDCommand {
+public class AlignToScoringLocationCommand extends EntechCommandBase {
 
+    protected final DriveSubsystem drive;
+    protected final PIDController pid;
+    protected final Supplier<DriveInput> operatorInput;
+    protected StoppingCounter counter;
     private ScoringLocationSupplier scoringLocationSupplier;
     private EstimatedPoseSupplier currentPoseSupplier;
     private AlignmentCalculator alignCalculator = new AlignmentCalculator();
@@ -37,9 +43,17 @@ public class AlignToScoringLocationCommand extends BaseDrivePIDCommand {
     		Supplier<DriveInput> operatorInput, 
     		ScoringLocationSupplier scoringLocationSupplier, 
     		EstimatedPoseSupplier currentPoseSupplier) {
-        super(drive,operatorInput);
+        super(drive);
+        this.drive = drive;
+        this.operatorInput = operatorInput;        
         this.scoringLocationSupplier = scoringLocationSupplier;
         this.currentPoseSupplier = currentPoseSupplier;
+        
+        pid = new PIDController(P_GAIN, I_GAIN, D_GAIN);
+        pid.enableContinuousInput(-180, 180);
+        pid.setTolerance(ANGLE_TOLERANCE);
+        counter = new StoppingCounter("PIDDriveCommand",STOP_COUNT);
+        
     }
 
     @Override
@@ -73,4 +87,19 @@ public class AlignToScoringLocationCommand extends BaseDrivePIDCommand {
         di.setRotation(calcValue);
         drive.drive(di );
     }
+    
+    @Override
+    public void end(boolean interrupted) {
+        drive.brake();
+    }
+
+    @Override
+    public boolean isFinished() {
+    	return counter.isFinished(pid.atSetpoint());    	
+    }
+
+    @Override
+    public boolean runsWhenDisabled() {
+        return false;
+    }    
 }
