+++ src/main/java/frc/robot/commands/DriveCommandConstants.java
@@ -1,0 +1,10 @@
+package frc.robot.commands;
+
+public class DriveCommandConstants {
+	public static final double P_GAIN = 0.0865;
+	public static final double I_GAIN = 0.5;
+    public static final double D_GAIN = 0.0075;
+    public static final double ANGLE_TOLERANCE = 1;
+    public static final double SPEED_LIMIT = 0.75;
+    public static final int STOP_COUNT = 4;
+}
