+++ src/main/java/frc/robot/subsystems/ArmSubsystem.java
@@ -5,7 +5,6 @@ import com.revrobotics.CANSparkMaxLowLevel.MotorType;
 import com.revrobotics.SparkMaxLimitSwitch.Type;
 
 import edu.wpi.first.util.sendable.SendableBuilder;
-import frc.robot.Robot;
 import frc.robot.RobotConstants;
 
 /**
@@ -24,7 +23,7 @@ public class ArmSubsystem extends EntechSubsystem{
   
   @Override
   public void initialize() {
-	if ( Robot.isReal() ) {
+	if ( enabled ) {
 	    elbowMotor = new CANSparkMax(RobotConstants.ARM.ELBOW_MOTOR_ID, MotorType.kBrushed);
 	    telescopeMotor = new CANSparkMax(RobotConstants.ARM.TELESCOPE_MOTOR_ID, MotorType.kBrushed);
 
@@ -36,7 +35,7 @@ public class ArmSubsystem extends EntechSubsystem{
 
   @Override
   public void initSendable(SendableBuilder builder) {
-	  if ( Robot.isReal() ) {
+	  if ( enabled ) {
 	      builder.setSmartDashboardType(getName());
 	      builder.addDoubleProperty("Elbow Motor", () -> { return elbowMotor.get(); }, null);
 	      builder.addDoubleProperty("Telescope Motor", () -> { return telescopeMotor.get(); }, null);		  
@@ -45,7 +44,9 @@ public class ArmSubsystem extends EntechSubsystem{
 
   @Override
   public void periodic() {
-    
+     if (enabled ) {
+    	 
+     }
   }
 
   @Override
@@ -54,10 +55,21 @@ public class ArmSubsystem extends EntechSubsystem{
   }
 
   private boolean isAtLowerElbowLimit() {
-    return elbowMotor.getReverseLimitSwitch(Type.kNormallyOpen).isPressed();
+	  if ( enabled ) {
+		  return elbowMotor.getReverseLimitSwitch(Type.kNormallyOpen).isPressed();
+	  }
+	  else {
+		  return false;
+	  }
+    
   }
 
   private boolean isAtFarTelescopeLimit() {
-    return telescopeMotor.getReverseLimitSwitch(Type.kNormallyOpen).isPressed();
+	  if (enabled) {
+		  return telescopeMotor.getReverseLimitSwitch(Type.kNormallyOpen).isPressed();
+	  }
+	  else {
+		  return false;
+	  }
   }
 }
